add unique, update strange point

+++ src/routes/charInven.router.js
@@ -34,16 +34,6 @@ router.post("/char/:charId/inventory/acquire", async (req, res, next) => {
   const { itemCode } = req.body;
 
   try {
-    const item = await prisma.items.findUnique({
-      where: {
-        itemCode: +itemCode,
-      },
-    });
-
-    if (!item) {
-      return res.status(404).json({ message: "아이템 존재 하지 않음" });
-    }
-
     const character = await prisma.character.findUnique({
       where: {
         charId: +charId,
@@ -54,16 +44,57 @@ router.post("/char/:charId/inventory/acquire", async (req, res, next) => {
       return res.status(404).json({ message: "캐릭터 찾을 수 없음" });
     }
 
-    const inventoryItem = await prisma.inventoryItem.create({
-      data: {
-        charId: +charId,
+    const isExistItem = await prisma.items.findUnique({
+      where: {
         itemCode: +itemCode,
       },
     });
 
-    return res
-      .status(200)
-      .json({ message: "아이템 획득", data: inventoryItem });
+    if (!isExistItem) {
+      return res.status(404).json({
+        message: `${isExistItem.itemName}이란 아이템은 존재 하지 않아`,
+      });
+    }
+
+    const existInvenItem = await prisma.charInven.findUnique({
+      where: {
+        charInven: {
+          charId: +charId,
+          itemCode: +itemCode,
+        },
+      },
+    });
+
+    if (existInvenItem) {
+      const updatedInventoryItem = await prisma.charInven.update({
+        where: {
+          charInvenId: existInvenItem.charInvenId,
+        },
+        data: {
+          quantity: {
+            increment: 1,
+          },
+        },
+      });
+
+      return res.status(200).json({
+        message: "아이템 수량 증가 완료",
+        data: updatedInventoryItem,
+      });
+    } else {
+      const newInvenItem = await prisma.charInven.create({
+        data: {
+          charId: +charId,
+          itemCode: +itemCode,
+          quantity: 1,
+        },
+      });
+
+      return res.status(200).json({
+        message: "아이템 획득 완료",
+        data: newInvenItem,
+      });
+    }
   } catch (error) {
     console.error(error);
     return res.status(400).json({ message: "에러", error });
@@ -75,7 +106,7 @@ router.delete("/char/:charId/inventory/:itemCode", async (req, res, next) => {
   const { charId, itemCode } = req.params;
 
   try {
-    const delItem = await prisma.charInven.findFirst({
+    const delItem = await prisma.charInven.findUnique({
       where: {
         charId: +charId,
         itemCode: +itemCode,
@@ -88,12 +119,22 @@ router.delete("/char/:charId/inventory/:itemCode", async (req, res, next) => {
       });
     }
 
-    await prisma.charInven.delete({
-      where: {
-        itemCode: +itemCode,
-        charId: +charId,
-      },
-    });
+    if (delItem.quantity > 1) {
+      await prisma.charInven.update({
+        where: {
+          charInvenId: delItem.charId,
+        },
+        data: {
+          quantity: quantity - 1,
+        },
+      });
+    } else {
+      await prisma.charInven.delete({
+        where: {
+          charInvenId: delItem.charInvenId,
+        },
+      });
+    }
 
     return res.status(200).json({ message: "아이템 삭제 완료" });
   } catch (error) {
